+++ src/usercloudssdk/asyncclient.py
@@ -11,7 +11,7 @@ import jwt
 
 from . import ucjson
 from .client_helpers import _SDK_VERSION, _id_from_identical_conflict, _read_env
-from .constants import _JSON_CONTENT_TYPE, AUTHN_TYPE_PASSWORD
+from .constants import _JSON_CONTENT_TYPE, AuthnType, Region
 from .errors import UserCloudsSDKError
 from .models import (
     Accessor,
@@ -84,8 +84,10 @@ class AsyncClient:
         self,
         id: uuid.UUID | None = None,
         organization_id: uuid.UUID | None = None,
-        region: str | None = None,
+        region: str | Region | None = None,
     ) -> uuid.UUID:
+        if isinstance(region, Region):
+            region = region.value
         body = {
             "id": id,
             "organization_id": organization_id,
@@ -100,12 +102,14 @@ class AsyncClient:
         password: str,
         id: uuid.UUID | None = None,
         organization_id: uuid.UUID | None = None,
-        region: str | None = None,
+        region: str | Region | None = None,
     ) -> uuid.UUID:
+        if isinstance(region, Region):
+            region = region.value
         body = {
             "username": username,
             "password": password,
-            "authn_type": AUTHN_TYPE_PASSWORD,
+            "authn_type": AuthnType.PASSWORD.value,
             "id": id,
             "organization_id": organization_id,
             "region": region,
@@ -166,8 +170,10 @@ class AsyncClient:
         row_data: dict,
         id: uuid.UUID | None = None,
         organization_id: uuid.UUID | None = None,
-        region: str | None = None,
+        region: str | Region | None = None,
     ) -> uuid.UUID:
+        if isinstance(region, Region):
+            region = region.value
         body = {
             "mutator_id": mutator_id,
             "context": context,
@@ -1006,7 +1012,7 @@ class AsyncClient:
 
     async def _prep_json_data_async(
         self, json_data: dict | str | None
-    ) -> tuple(dict, str | None):
+    ) -> tuple[dict, str | None]:
         await self._refresh_access_token_if_needed_async()
         headers = self._get_headers()
         content = None