+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Add GetConsentedPurposesForUser method to the Python SDK Client.
 - Add `__repr__` and `__str__` methods to some models to improve DX.
 - Add `to_dict()` method to the Address class to make it easier to serialize it into user row when creating or modifying a user.
+- Add Region enum and move a lot of other constants into enums (AuthnType, ColumnIndexType, DataLifeCycleState, DataType, PolicyType, TransformType).
+- Add `data_life_cycle_state` to Accessor model.
 
 ## 1.4.0 - 02-02-2024